+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 
 - keep the order of contacts when calling getContactsByIds #4651, #4652
 
+### Fixed
+- fix unread count on "jump down" button not clearing when all messages are read #4648
+- keep the order of contacts when calling getContactsByIds #4651
+- improve accessibility #4655, #4656, #4661, #4662
+
 <a id="1_54_0"></a>
 
 ## [1.54.0] - 2025-02-15